move cache expiry to a sweeper
git-svn-id: https://svn.radicaldesigns.org/daysofaction@342 8f85c81b-0f21-0410-93f9-e2b37b76a6fe

+++ trunk/app/models/calendar.rb
@@ -12,39 +12,17 @@ class Calendar < ActiveRecord::Base
   end
 
   def self.clear_deleted_events
-    find_deleted_events
-    @@deleted_events.each do |e|
-      e.destroy
-      FileUtils.rm(File.join(ActionController::Base.page_cache_directory,"events/show/#{e.id}.html")) rescue Errno::ENOENT
-      FileUtils.rm(File.join(ActionController::Base.fragment_cache_store.cache_path, "events/_report/event_#{e.id}_list_item.cache")) rescue Errno::ENOENT
+    events = Event.find(:all, :conditions => "service_foreign_key != 0")
+    events.each do |e|
+      e.destroy if e.service_foreign_key && !DemocracyInActionEvent.find(e.service_foreign_key)
     end
-    unless @@deleted_events.empty?
-      FileUtils.rm(File.join(ActionController::Base.page_cache_directory,'index.html')) rescue Errno::ENOENT
-      FileUtils.rm(File.join(ActionController::Base.page_cache_directory,'events','total.html')) rescue Errno::ENOENT
-      FileUtils.rm(File.join(ActionController::Base.page_cache_directory,'events','flashmap.xml')) rescue Errno::ENOENT
-      @@deleted_events.collect {|e| e.state}.compact.uniq.each do |s|
-        FileUtils.rm(File.join(ActionController::Base.page_cache_directory,'events','search','state',"#{s}.html")) rescue Errno::ENOENT
-      end
-    RAILS_DEFAULT_LOGGER.info("Caches fully swept after deleting #{@@deleted_events.length} events")
-    end
-  end
-
-  def self.find_deleted_events
-    @@all_events = Event.find(:all)
-    opts = YAML.load_file(File.join(RAILS_ROOT,'config','democracyinaction-config.yml'))
-    require 'DIA_API_Simple'
-    api = DIA_API_Simple.new opts
-    @@all_events.each do |e|
-      dia_event = api.get 'event', e.service_foreign_key
-      @@deleted_events << e if dia_event.empty?
-    end
-    @@deleted_events
   end
 
   DiaLoadResult = Struct.new(:imported, :unknown, :inaccurate)
   def self.load_from_dia(id, *args)
     cal = find(id)
     return unless cal
+    #TODO: use DemocracyInActionEvent
     options = args.last.is_a?(Hash) ? args.pop : {}
     opts = YAML.load_file(File.join(RAILS_ROOT,'config','democracyinaction-config.yml'))
     require 'DIA_API_Simple'
@@ -60,9 +38,6 @@ class Calendar < ActiveRecord::Base
       gg = GoogleGeocode::Accuracy.new application_id
     end
 
-    states = []
-    rm_index = false
-    rm_flashmap = false
     events.each do |e|
       my_event = Event.find_or_initialize_by_service_foreign_key(e['event_KEY'])
       next if !my_event.new_record? &&
@@ -76,11 +51,6 @@ class Calendar < ActiveRecord::Base
         my_event.end          == e['End'].to_time(:local) &&
         my_event.directions   == e['Directions']
 
-      rm_index ||= my_event.new_record?
-      rm_flashmap ||= my_event.new_record? || my_event.postal_code != e['Zip'] ||
-        my_event.city != e['City'] || my_event.state != e['State'] || 
-        my_event.name != e['Event_Name']
-
       my_event.calendar_id = id
       my_event.name = e['Event_Name']
       my_event.description = e['Description']
@@ -91,8 +61,6 @@ class Calendar < ActiveRecord::Base
       my_event.start = e['Start']
       my_event.end = e['End']
       my_event.directions = e['Directions']
-
-      states << my_event.state
 
       if gg
         begin
@@ -115,20 +83,8 @@ class Calendar < ActiveRecord::Base
         UserMailer.deliver_invalid(my_event, err)
         my_event.save(false)
       end
-      FileUtils.rm(File.join(ActionController::Base.page_cache_directory,"events/show/#{my_event.id}.html")) rescue Errno::ENOENT
-      FileUtils.rm(File.join(ActionController::Base.fragment_cache_store.cache_path, "events/_report/event_#{my_event.id}_list_item.cache")) rescue Errno::ENOENT
     end
-
     result.imported = events.length
-
-    states.compact.uniq.each do |s|
-      FileUtils.rm(File.join(ActionController::Base.page_cache_directory,'events','search','state',"#{s}.html")) rescue Errno::ENOENT
-    end
-
-    FileUtils.rm(File.join(ActionController::Base.page_cache_directory,'index.html')) rescue Errno::ENOENT if rm_index
-    FileUtils.rm(File.join(ActionController::Base.page_cache_directory,'events','total.html')) rescue Errno::ENOENT if rm_index
-    FileUtils.rm(File.join(ActionController::Base.page_cache_directory,'events','flashmap.xml')) rescue Errno::ENOENT if rm_flashmap
-    RAILS_DEFAULT_LOGGER.info("Caches fully swept after adding #{events.length} events") if rm_index || rm_flashmap || !states.empty?
 
     return result
   end
